requirements.txt
requests
openai
PySocks